Changed undo/redo icons, other icons incoming! Hopefully. I'm digging through files for the designer icons.
Also started a slight refactor of the help dialog to try to select
qtreeview items from the help menu.

On the font issue, it's not even working correctly in QT's example code,
but I have yet to test it with a CSS sheet, which may help?

Also added a example input code field, as well as button in the help
dialog (+ helper function) but haven't implemented it in the main window
yet.

Images are also working in the help system! Woohoo! (See Debugging
Programs for an example)

Help dialog also opens with the submenus expanded by default now, which
makes more sense I think.


git-svn-id: http://pep8-1.googlecode.com/svn/trunk@104 0efa5a84-20a4-11de-9909-f37ff7dbbc12