Save system update. Will now insert file names with extensions into the save dialog!
Still loose ends to tie up. Need to think on how to set the current
files when object code is saved.

Also, the open dialog has been improved, can only select .txt, .pep, and
.pepo files.

Need to find a way to save to a certain directory by platform.

Implemented the beginnings of the debug simulation (checking and
unchecking boxes in the CPU pane)


git-svn-id: http://pep8-1.googlecode.com/svn/trunk@141 0efa5a84-20a4-11de-9909-f37ff7dbbc12