Removed universal binary option from .pro file because we're going to use lipo to combine static builds.
The CPU pane now can get focus (by clicking anywhere on the pane), and
automatically gets focus when debugging begins. This allows us to
accept 'enter' keystrokes to single step.

We no longer resize the document width in the listing trace pane based
on the splitter size, but the code is there if we want to bring it back
later.


git-svn-id: http://pep8-1.googlecode.com/svn/trunk@316 0efa5a84-20a4-11de-9909-f37ff7dbbc12