Added a temporary scroll-to screenshot for the memory dump pane documentation. We should also now be able to interrupt execution inside a trap in the operating system when not tracing traps. Help dialog tree widget font size fix(?) - hasn't been tested under ubuntu or windows yet.
I have an idea of how to fix the display bug when running in the memory
trace pane, but I'm going to sleep on it (we may need to cache the items
before adding them to the scene, which we were doing before, but removed
that feature because we thought it was superflous due to how Qt handles
items. Apparently not?).


git-svn-id: http://pep8-1.googlecode.com/svn/trunk@395 0efa5a84-20a4-11de-9909-f37ff7dbbc12