Extends the CPU pane to display ASCII forms
Display the values in an ASCII form for the accumulator, the index register
and the operand. Each byte of the word is displayed independently to
easily spot them.

In case of non printable value, the symbolic name of the ASCII character
is displayed (eg. "NUL").

The SP, PC and OPCODE does not have this extended ASCII display since it
makes little sense.

This additional information may be useful for students for two reasons:
* It helps to see what character is actually stored in registers or in
  operands without having to perform conversion or to look up in a ASCII
  table
* It helps to distinguish characters stored in the low byte of a word
  between those stored in the high byte of a word (because of LDr instead
of a LDBYTEr for instance)

The following programs highlight some character manipulations. This patch
could help students to figure out what is going on during a trace session.

; Misuse of LDr vs LDBYTEr
         LDA     chr,d       ; 'a' is displayed in the first ASCII cell
         LDBYTEX chr,d       ; 'a' is displayed is the second cell
         STOP
chr:     .BYTE   'a'
data:    .BYTE   0x5A        ; unrelated other data
         .END

; Bad adressing
         LDBYTEA chr,i       ; no 'a' in the cell, why?
         STOP
chr:     .BYTE   'a'
         .END